moving document List to a separate reducer (#4169)
moving document List to a separate reducer
