+++ client/app/reader/Documents/DocumentsActions.js
@@ -3,7 +3,8 @@ import * as Constants from './actionTypes';
 import ApiUtil from '../../util/ApiUtil';
 import { hideErrorMessage, showErrorMessage, updateFilteredIdsAndDocs } from '../commonActions';
 import { categoryFieldNameOfCategoryName } from '../utils';
-import { collectAllTags, handleSetLastRead } from '../PdfViewer/PdfViewerActions';
+import { collectAllTags } from '../PdfViewer/PdfViewerActions';
+import { handleSetLastRead } from '../DocumentList/DocumentListActions';
 import { setViewedAssignment } from '../CaseSelect/CaseSelectActions';
 import { setLoadedVacolsId } from '../LoadingScreen/LoadingScreenActions';
 import { CATEGORIES, ENDPOINT_NAMES } from '../analytics';
