+++ client/app/reader/Documents/actionTypes.js
@@ -5,8 +5,8 @@ export const ROTATE_PDF_DOCUMENT = 'ROTATE_PDF_DOCUMENT';
 export const SELECT_CURRENT_VIEWER_PDF = 'SELECT_CURRENT_VIEWER_PDF';
 export const TOGGLE_DOCUMENT_CATEGORY_FAIL = 'TOGGLE_DOCUMENT_CATEGORY_FAIL';
 export const TOGGLE_DOCUMENT_CATEGORY = 'TOGGLE_DOCUMENT_CATEGORY';
-export const TOGGLE_COMMENT_LIST = 'TOGGLE_COMMENT_LIST';
 export const SET_VIEWING_DOCUMENTS_OR_COMMENTS = 'SET_VIEWING_DOCUMENTS_OR_COMMENTS';
+export const TOGGLE_COMMENT_LIST = 'TOGGLE_COMMENT_LIST';
 
 // Tags
 export const REQUEST_NEW_TAG_CREATION_SUCCESS = 'REQUEST_NEW_TAG_CREATION_SUCCESS';
