+++ client/app/reader/PdfUI.jsx
@@ -285,7 +285,7 @@ const mapStateToProps = (state, props) => {
   const numPages = pdfDocument ? pdfDocument.pdfInfo.numPages : null;
 
   return {
-    ..._.pick(state.readerReducer.ui, 'filteredDocIds'),
+    ..._.pick(state.documentList, 'filteredDocIds'),
     docListIsFiltered: docListIsFiltered(state),
     loadError: state.readerReducer.documentErrors[props.doc.content_url],
     isPlacingAnnotation: state.annotationLayer.isPlacingAnnotation,
