+++ client/app/reader/constants.js
@@ -2,15 +2,9 @@ import { CategoryIcon } from '../components/RenderFunctions';
 import React from 'react';
 
 // actions
-export const RECEIVE_MANIFESTS = 'RECEIVE_MANIFESTS';
 export const RECEIVE_ANNOTATIONS = 'RECEIVE_ANNOTATIONS';
 export const RECEIVE_APPEAL_DETAILS = 'RECEIVE_APPEAL_DETAILS';
 export const RECEIVE_APPEAL_DETAILS_FAILURE = 'RECEIVE_APPEAL_DETAILS_FAILURE';
-export const TOGGLE_FILTER_DROPDOWN = 'TOGGLE_FILTER_DROPDOWN';
-export const SET_CATEGORY_FILTER = 'SET_CATEGORY_FILTER';
-export const SET_TAG_FILTER = 'SET_TAG_FILTER';
-export const CLEAR_CATEGORY_FILTER = 'CLEAR_CATEGORY_FILTER';
-export const CLEAR_TAG_FILTER = 'CLEAR_TAG_FILTER';
 export const SHOW_NEXT_PDF = 'SHOW_NEXT_PDF';
 export const SHOW_PREV_PDF = 'SHOW_PREV_PDF';
 export const SCROLL_TO_COMMENT = 'SCROLL_TO_COMMENT';
@@ -18,14 +12,8 @@ export const TOGGLE_PDF_SIDEBAR = 'TOGGLE_PDF_SIDEBAR';
 export const TOGGLE_SEARCH_BAR = 'TOGGLE_SEARCH_BAR';
 export const SHOW_SEARCH_BAR = 'SHOW_SEARCH_BAR';
 export const HIDE_SEARCH_BAR = 'HIDE_SEARCH_BAR';
-export const LAST_READ_DOCUMENT = 'LAST_READ_DOCUMENT';
 export const SCROLL_TO_SIDEBAR_COMMENT = 'SCROLL_TO_SIDEBAR_COMMENT';
 export const COLLECT_ALL_TAGS_FOR_OPTIONS = 'COLLECT_ALL_TAGS_FOR_OPTIONS';
-export const SET_SORT = 'SET_SORT';
-export const SET_SEARCH = 'SET_SEARCH';
-export const CLEAR_ALL_FILTERS = 'CLEAR_ALL_FILTERS';
-export const CLEAR_ALL_SEARCH = 'CLEAR_ALL_SEARCH';
-export const SET_DOC_LIST_SCROLL_POSITION = 'SET_DOC_LIST_SCROLL_POSITION';
 export const SET_DOC_SCROLL_POSITION = 'SET_DOC_SCROLL_POSITION';
 export const JUMP_TO_PAGE = 'JUMP_TO_PAGE';
 export const RESET_JUMP_TO_PAGE = 'RESET_JUMP_TO_PAGE';
@@ -42,17 +30,10 @@ export const SET_SEARCH_INDEX_TO_HIGHLIGHT = 'SET_SEARCH_INDEX_TO_HIGHLIGHT';
 export const SET_DOCUMENT_LOAD_ERROR = 'SET_DOCUMENT_LOAD_ERROR';
 export const CLEAR_DOCUMENT_LOAD_ERROR = 'CLEAR_DOCUMENT_LOAD_ERROR';
 export const SET_LOADED_APPEAL_ID = 'SET_LOADED_APPEAL_ID';
-export const UPDATE_FILTERED_RESULTS = 'UPDATE_FILTERED_RESULTS';
 
 export const SET_OPENED_ACCORDION_SECTIONS = 'SET_OPENED_ACCORDION_SECTIONS';
-export const SET_VIEWING_DOCUMENTS_OR_COMMENTS = 'SET_VIEWING_DOCUMENTS_OR_COMMENTS';
 
 export const COMMENT_ACCORDION_KEY = 'Comments';
-
-export const DOCUMENTS_OR_COMMENTS_ENUM = {
-  DOCUMENTS: 'documents',
-  COMMENTS: 'comments'
-};
 
 // If we used CSS in JS, we wouldn't have to keep this value in sync with the CSS in a brittle way.
 export const ANNOTATION_ICON_SIDE_LENGTH = 40;
